test(cli): put the play module's tests in the merge gate
`play` is off by default, and `mod play` was gated on it, so the whole
module including its `mod tests` compiled out of every default build.
`just check`, `just test`, and check.yml are default features only, so
nothing a reviewer sees ran those tests; the only thing that did was
nightly's `just rs features`, behind two workspace builds that can fail
first.

Split `play.rs` into a directory module. The logic compiles
unconditionally (`args`, `layout`, `playback`, `source`, `timeline`) and
only the winit/wgpu/cpal event loop stays behind the feature (`media`,
`window`). No new dependency and no new CI job: the twelve tests now run
in the default `just test`.
