+++ rs/CLAUDE.md
@@ -144,7 +144,9 @@ Then `Config::load()?` (initializes tracing), build clients/servers via `.init()
 
 - `just check` lints and compiles the crates your branch changed plus every crate depending on them; `just test` runs their tests; `just fix` auto-fixes formatting/lint over the same set (`just rs _select` does the selection, via `cargo metadata`). `just check-all` / `just test all` / `just fix-all` cover every default member. `just rs test -p <crate>` (or `cargo nextest run -p <crate>`) for one crate.
 
-- **`check` compiles default features only, and so does CI.** The permutations moved to `just rs features` (nightly): `--all-features` costs a full extra workspace compile that shares almost no artifacts with the default one (measured at ~6 minutes on top of an already-warm tree), and `--no-default-features` is a third distinct feature set that shares nothing with either. `features` is the only thing that compiles moq-cli's `play`/`capture`, moq-audio's capture backend, quiche, and jemalloc, so a break in those lands on `main` and surfaces nightly rather than in review. `just rs audit` (cargo-deny) is nightly for the same workflow reason: an advisory is published without this repo changing.
+- **`check` compiles default features only, and so does CI.** The permutations moved to `just rs features` (nightly): `--all-features` costs a full extra workspace compile that shares almost no artifacts with the default one (measured at ~6 minutes on top of an already-warm tree), and `--no-default-features` is a third distinct feature set that shares nothing with either. `features` is the only thing that compiles moq-cli's `play` event loop and `capture`, moq-audio's capture backend, quiche, and jemalloc, so a break in those lands on `main` and surfaces nightly rather than in review. `just rs audit` (cargo-deny) is nightly for the same workflow reason: an advisory is published without this repo changing.
+
+- **An off-by-default feature gate hides that code's tests from the merge gate, so keep it around the dependency and not around the logic.** `check` and `test` are default features only, so a `#[cfg(feature = ...)]` module's `mod tests` is never compiled in review, and `just rs features` is a once-a-day run sitting behind two other workspace builds that can fail first. A test that cannot pass therefore reads green. `moq-cli`'s `play` is the worked example: the argument validation, rendition selection, and media timing live in `play/{args,layout,playback,source,timeline}.rs` and compile either way, while only the winit/wgpu/cpal event loop (`play/{media,window}.rs`) is behind the feature. `capture` and `transcode` are gated the same way and have no tests yet; theirs belong on the default-compiled side too.
 
 - **`check` runs no `cargo check` pass.** Clippy is a superset of it, and the two use different rustc wrappers, so running both compiles the workspace twice for one set of errors.
